get rid of set_version
