+++ packages/kit/src/runtime/server/index.js
@@ -1,5 +1,5 @@
 import { respond } from './respond.js';
-import { set_private_env, set_public_env } from '../shared.js';
+import { set_private_env, set_public_env } from '../shared-server.js';
 import { options, get_hooks } from '__SERVER__/internal.js';
 
 export class Server {
